CUA_approach/requirements.txt
openai>=1.6.0
python-dotenv>=1.0.0
playwright>=1.40.0
pytesseract>=0.3.10
Pillow>=10.0.0
